+++ documentation/docs/src/lib/components/Form/Select/Select.stories.svelte
@@ -141,6 +141,9 @@
     customTriggerIcon: {
       table: { category: 'API', subcategory: 'Slots' },
     },
+    customListItemContent: {
+      table: { category: 'API', subcategory: 'Slots' },
+    },
   };
 
   // ------------------------------
@@ -205,6 +208,17 @@
   name="No Trigger Button"
   args={{ placeholder: 'Search a number', searchable: true, showTriggerButton: false }}
 />
+
+<Story name="Custom Popup Custom List Item Content" asChild>
+  <Select {options}>
+    {#snippet customListItemContent({ option, selectedValue })}
+      {#if selectedValue === option.value}
+        ✅
+      {/if}
+      {option.label} ✨
+    {/snippet}
+  </Select>
+</Story>
 
 <Story name="Custom Trigger Icon" asChild>
   <Select {options}>