+++ packages/dodo-ui-bits/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flightlesslabs/dodo-ui-bits
 
+## 0.10.6
+
+### Patch Changes
+
+- Select customListItemContent added
+
 ## 0.10.5
 
 ### Patch Changes